backend/pyproject.toml
[project]
name = "cadence"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.1.0",
]

[tool.setuptools.packages.find]
include = ["app*", "evals*"]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.13.0",
]
eval = [
    "google-genai>=1.0.0",
]
